% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastshap.R
\name{ForcePlot.explain}
\alias{ForcePlot.explain}
\title{Additive force plots for fastshap explanation objects}
\usage{
\method{ForcePlot}{explain}(shaps, features, baseValue = NULL, ...)
}
\arguments{
\item{shaps}{An \code{explain} object created with \code{\link[fastshap:explain]{fastshap::explain()}}}

\item{features}{Data used for model training or new data that was explained}

\item{baseValue}{Optional baseline for predictions (e.g. mean of all
predictions)}

\item{...}{Additional parameters for \code{\link[=AdditiveForcePlot]{AdditiveForcePlot()}} or
\code{\link[=AdditiveForceArrayPlot]{AdditiveForceArrayPlot()}}}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Additive force plots for fastshap explanation objects
}
\examples{
if (interactive() & require("fastshap")) {
   data(mtcars)
   mtcars.ppr <- ppr(mpg ~ ., data = mtcars, terms = 1)
   # Compute approximate Shapley values using 10 Monte Carlo simulations
   set.seed(101)  # for reproducibility
   shap <- explain(fit, X = subset(mtcars, select = -mpg), nsim = 10,
               pred_wrapper = predict)
   # Visualize first explanation
   preds <- predict(mtcars.ppr, newdata = mtcars)
   ForcePlot(shap[1,], mtcars[1,], mean(preds))
}
}
