% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ForcePlot}
\alias{ForcePlot}
\title{ForcePlots for compatible objects}
\usage{
ForcePlot(...)
}
\arguments{
\item{...}{Parameters for the specific functions}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Generic method for \code{shapper}, \code{shapr} and \code{fastshap} objects.
See \link{ForcePlot.individual_variable_effect},
\link{ForcePlot.shapr} and \link{ForcePlot.explain} for
specific documentation.
}
