% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapr.R
\name{ForcePlot.shapr}
\alias{ForcePlot.shapr}
\title{Display additive force plots for shapr explanation objecs}
\usage{
ForcePlot.shapr(explanation, i = NULL, ...)
}
\arguments{
\item{explanation}{A \code{shapr} object, created with the \code{\link[shapr:explain]{shapr::explain()}}
method}

\item{i}{Index for single object explanation}

\item{...}{Additional parameters to \code{\link[=AdditiveForcePlot]{AdditiveForcePlot()}} or
\code{\link[=AdditiveForceArrayPlot]{AdditiveForceArrayPlot()}}}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Display additive force plots for shapr explanation objecs
}
