% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{SimpleListPlot}
\alias{SimpleListPlot}
\title{Display simple feature effect plot}
\usage{
SimpleListPlot(
  features,
  featureNames = NULL,
  plot_cmap = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{features}{A named list of features with \code{effect} and optional
\code{name}}

\item{featureNames}{Optional named list or vector of feature names}

\item{plot_cmap}{Optional color map to pass to component or list of two valid
web color names or hex codes. Valid color maps are: RdBu, GnPR, CyPU, PkYg,
DrDb, LpLb, YlDp and OrId.}

\item{width}{Optional widget width}

\item{height}{Optional widget height}
}
\description{
This function uses the SimpleListVisualizer component from the shapjs package
}
\examples{
feature_1 <- list(name = "A", effect = 0.5)
feature_2 <- list(name = "B", effect = -1)
features <- list(feature_1, feature_2)
SimpleListPlot(features)
}
