% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{AdditiveForceArrayPlot}
\alias{AdditiveForceArrayPlot}
\title{Additive force plots for array of explanations}
\usage{
AdditiveForceArrayPlot(
  baseValue,
  explanations,
  featureNames,
  outNames = "",
  link = c("identity", "logit"),
  plot_cmap = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{baseValue}{Baseline for predictions}

\item{explanations}{List of explanations containing outValue (predicted
value), simIndex (similarity index) and features list of effects and values.}

\item{featureNames}{Optional named list of feature names}

\item{outNames}{Optional list of the target variable names}

\item{link}{Optional link function to use (identity is default)}

\item{plot_cmap}{Optional color map to pass to component}

\item{width}{Optional widget width}

\item{height}{Optional optional widget height}
}
\description{
Create a force plot widget for multiple explanations
}
