% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{AdditiveForcePlot}
\alias{AdditiveForcePlot}
\title{Single additive force plot}
\usage{
AdditiveForcePlot(
  baseValue,
  features,
  featureNames = NULL,
  outNames = NULL,
  link = c("identity", "logit"),
  plot_cmap = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{baseValue}{Baseline for predictions}

\item{features}{List of features effects and values. This should be a list of
lists where each sublist has names effect and value. See example.}

\item{featureNames}{Optional named list of feature names}

\item{outNames}{Optional list of the target variable names}

\item{link}{Optional link function to use (identity is default)}

\item{plot_cmap}{Optional color map to pass to component}

\item{width}{Optional widget width}

\item{height}{Optional optional widget height}
}
\description{
Create a force plot widget for a single explanation
}
