% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{ForcePlot}
\alias{ForcePlot}
\title{ForcePlots for compatible objects}
\usage{
ForcePlot(...)
}
\arguments{
\item{...}{Parameters for the specific functions}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Method for \code{individual_variable_effect} and \code{shapr}
objects. See \code{\link[=ForcePlot.individual_variable_effect]{ForcePlot.individual_variable_effect()}} and
\code{\link[=ForcePlot.shapr]{ForcePlot.shapr()}} for further documentation.
}
