% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapr.R
\name{ForcePlot.shapr}
\alias{ForcePlot.shapr}
\title{Additive force plots for shapr explanation objects}
\usage{
\method{ForcePlot}{shapr}(explanation, i = NULL, ...)
}
\arguments{
\item{explanation}{A \code{shapr} object created with \code{\link[shapr:explain]{shapr::explain()}}}

\item{i}{Index for single object explanation}

\item{...}{Additional parameters for \code{\link[=AdditiveForcePlot]{AdditiveForcePlot()}} or
\code{\link[=AdditiveForceArrayPlot]{AdditiveForceArrayPlot()}}}
}
\value{
An object of class \code{htmlwidget}
}
\description{
Additive force plots for shapr explanation objects
}
