% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{SimpleListPlot}
\alias{SimpleListPlot}
\title{Display simple feature plot}
\usage{
SimpleListPlot(
  features,
  featureNames = NULL,
  plot_cmap = NULL,
  width = NULL,
  height = NULL
)
}
\arguments{
\item{features}{A named list of features with effects}

\item{featureNames}{Optional named list matching indices with names}

\item{plot_cmap}{Optional color map to pass to component}

\item{width}{Optional widget width}

\item{height}{Optional widget height}
}
\description{
This function uses the SimpleListVisualizer component from the shapjs package
}
